requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
ReportLab==4.0.7
email-validator==2.0.0
PyJWT==2.8.0
alembic==1.13.0
Flask-Login==0.6.3
